+++ js/main.js
@@ -2,4 +2,3 @@ import { getRandomObjects } from './data';
 
 getRandomObjects(25);
 
-
